Added Hit Decals
When bullets hit the player while shielded, a hit decal is instantiated.
