+++ Assets/Editor/CrossPlatformInput.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 41e4f29e5dee9ec48a2538955ef1de71
+folderAsset: yes
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
