--- Assets/LevelChange.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 7d954cca5c8016344860409d8ecbb0af
-timeCreated: 1455240929
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
