+++ Assets/Materials/New Material.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2c9344ff083860c4aa8eb70b0405bec6
+timeCreated: 1455262266
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
