+++ Assets/Prefabs/Bullet.prefab
@@ -145,6 +145,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bb0a7de91fd123043ad2ada49a57377e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  decal: {fileID: 106938, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
   brackets: {fileID: 0}
 --- !u!114 &11420726
 MonoBehaviour:
@@ -334,6 +335,10 @@ Prefab:
       propertyPath: m_LocalScale.z
       value: 0.5
       objectReference: {fileID: 0}
+    - target: {fileID: 0}
+      propertyPath: decal
+      value: 
+      objectReference: {fileID: 106938, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 0}
   m_RootGameObject: {fileID: 181106}
