+++ Assets/Prefabs/Shield Decal.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7eef142717ab9324f9cdbd816be9677f
+timeCreated: 1455268381
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
