+++ Assets/Scenes/PlayerEditor.unity
@@ -351,8 +351,79 @@ Prefab:
       propertyPath: m_RootOrder
       value: 2
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 13691722, guid: b8c07257685632c429aeb349495d3c37, type: 2}
   m_ParentPrefab: {fileID: 100100000, guid: b8c07257685632c429aeb349495d3c37, type: 2}
+  m_IsPrefabParent: 0
+--- !u!1 &1356511764 stripped
+GameObject:
+  m_PrefabParentObject: {fileID: 136150, guid: b8c07257685632c429aeb349495d3c37, type: 2}
+  m_PrefabInternal: {fileID: 1356511763}
+--- !u!64 &1356511765
+MeshCollider:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 1356511764}
+  m_Material: {fileID: 0}
+  m_IsTrigger: 1
+  m_Enabled: 1
+  serializedVersion: 2
+  m_Convex: 1
+  m_Mesh: {fileID: 4300006, guid: 11fc3834e3194b94f85ad5cda52ad7ad, type: 3}
+--- !u!1001 &1800481495
+Prefab:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 464946, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
+      propertyPath: m_LocalPosition.x
+      value: -0.63310593
+      objectReference: {fileID: 0}
+    - target: {fileID: 464946, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
+      propertyPath: m_LocalPosition.y
+      value: 0.03515625
+      objectReference: {fileID: 0}
+    - target: {fileID: 464946, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
+      propertyPath: m_LocalPosition.z
+      value: -0.0337525
+      objectReference: {fileID: 0}
+    - target: {fileID: 464946, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 464946, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 464946, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 464946, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 464946, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
+      propertyPath: m_RootOrder
+      value: 8
+      objectReference: {fileID: 0}
+    - target: {fileID: 464946, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
+      propertyPath: m_LocalScale.x
+      value: 0.1
+      objectReference: {fileID: 0}
+    - target: {fileID: 464946, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
+      propertyPath: m_LocalScale.y
+      value: 0.1
+      objectReference: {fileID: 0}
+    - target: {fileID: 464946, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
+      propertyPath: m_LocalScale.z
+      value: 0.1
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_ParentPrefab: {fileID: 100100000, guid: 7eef142717ab9324f9cdbd816be9677f, type: 2}
   m_IsPrefabParent: 0
 --- !u!1001 &1938929607
 Prefab:
