+++ Assets/Scripts/Shield and HUD/ShieldDecalController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 1822970ef5f4f944ca6b96888cd15ec0
+timeCreated: 1455268436
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
